[svn] Substantial reorganization of the manifest and iso-related events primarily to move the events to the DiffMixin/FilesMixin models.  Major changes include -
* Moved the manifest out of iso.py since it's purpose (providing a
control list of files in the distribution) is general.
* Split the iso event into two events (pkgorder and iso-sets) with a
commmon parent (ISO).
* Added support for specifying a user provided pkgorder file in
configuration.  Updated schema and documentation files to
reflect this change. Removed previous command line --with-pkgorder
option.
* As mentioned above, moved all three events (manifest, pkgorder and
iso-sets) to DiffMixin.  Moved pkgorder to FilesMixin.  Laid the
groundwork in pkgorder and iso-sets to begin using a future version of
FilesMixin which supports file paths in addition to config elements.