[svn] Summary: Renamed software.py to download.py; svn required this in a separate step.
Note: Logging for download, gpgsign, and createrepo events is still a
work in progress. For example, at present if you have both checking and
signing enabled you will see the list of rpms fly by a total of 4 times.  We
should be able to collapse this to three: downloading, checking,
signing.