[svn] Warning - this commit is a result of svn's retarded insistence that you can't move a file a second time until after it has been committed.  Check out and use at your own risk
Summary: Initial commit of the concept of 'default-on', 'default-off' for
modules.  Created two subfolders of modules, 'core' and 'extensions', which
are to be used to contain modules according to the following rough heuristic:

 * core modules are modules that are required for dimsbuild to complete
   successfully (comps.py, pkglist.py) or those that are deemed important
   enough in the process as to be required for a distribution to be 'complete'
   (rpms/release.py, publish.py).  Core modules are 'default-on', meaning that
   unless specified otherwise, they are loaded and executed normally.
 * extension modules are modules that extend, modify, or otherwise manipulate
   the output of dimsbuild in a non-process critical way (files.py, iso.py).
   Extension modules are 'default-off', meaning that they have to be explicitly
   enabled in the config file or at the command line in order for them to
   run

Currently all modules are classfied as 'core'; this will be adjusted on a
module-by-module basis in the future.

Modules can be disabled/enabled in two ways.  First, any top-level element in
the config file is automatically enabled by default, unless the 'default'
attribute is set to False.  Modules that aren't explicitly listed in the config
file are treated according to their type; core modules are loaded, extension
modules are not.  Secondly, the command line options --enable-module and
--disable-module can be used to explicitly enable and disable modules.

Also created a third subfolder of modules, 'lib'.  This folder should contain
any and all shared library files for all modules.

All events are now listed in their config files as top level items.  Both the
code and schemas have been updated to reflect this.

All schemas modified slightly in an attempt to make them less vulnerable to
status changes (especially optional) and more parallel to the files they are
compared to.  Instead of indenting at each level of xml, indents only occur
when the underlying XML document being checked would indent.  This means that
the entire document doesn't have to be readjusted whenever an <optional> tag
is removed, for example.  It also makes the schema and the XML document itself
roughly parallel in structure.  Its primary disadvantage is it is somewhat
more difficult to read.  I've tried to counteract this with liberal use of
newlines, but if we don't like it, we can revert this change (as usual).