[svn] Summary: Renamed all modules so that they correspond to their config top-level element names.
WARNING: this makes some events have 'illegal' names as far as python is
concerned for direct importing (specifically, they contain the '-'
character).  However, the only time this is a problem is if you are
trying to directly import these modules using the import statement, which
doesn't happen in dimsbuild (and, incidentally, doesn't work anyway due
to the complex environment required.  If we determine for some reason
that we really don't like this restriction, I suggest converting the '-'
to '_' in the filenames only, and having dimsbuild handle the conversion
back to '-' internally.

Rationale behind '-' vs '_': '-' is easier to type, easier to see, and
more commonly used outside of programming circles than '_'.  '-' also
occurs much more commonly in xml document tag names.