repos: detect yum variables in repo config and raise error
Yum variables (e.g. "$releasever, $arch, $basearch, and $YUM0 - $YUM9)
are ambiguous in the distribution build context. For example, should
$releasever be the release number of a) the build machine, b) the
distribution being built, or c) the base repo?.  This change detects the
presence of yum variables and raises a ValueError directing the user to replace yum variables with fixed values.