Bug 79 : config-rpm fails if the element has no contents.
* lib/dimsbuild/modules/shared/rpms.py:
- the RpmBuildMixin._test_build() function is no longer needed,
because the event is not loaded if it is disabled taking care of
everything _test_build() used to.

* lib/dimsbuild/modules/{core,extensions}/rpms:
- not using RpmBuildMixin._test_build()

* lib/dimsbuild/modules/extensions/rpms/config-rpm.py:
- the ConfigRpmEvent had the _test_build() function that overrode the
one in RpmBuildMixin. This method returned true iff there were
elements in the <config-rpm/> element. This method has also been
deleted because if a user explicitly enables an extension module, it
should generate output.