RepoEventMixin: remove repo if enabled=[false] in repofile
Prior to this commit, dimsbuild ignored the enabled property in
provided yum .repo files.  With this commit, dimsbuild notices the
enabled property and removed the repo if it is disabled.

lib/dimsbuild/modules/shared/repos.py
* read_config() - during post-processing of newly created repo objects,
test if repo.has_key('enabled') and repo['enabled'] in BOOLEANS_FALSE,
if so, pop repo from repocontainer object and continue to next repo.