pkglist: potential fix to bug 95.
* lib/dimsbuild/modules/core/software/pkglist.py:
- in incremental depsolve's handler, I check to make sure that a
package that was required by some other package the last time, is
still the latest one this time. To accomplish this, I make use of the
DepSolver._provideToPkg() method, which can return None if the package
is not available. So, in the case when you change input repositories,
the package might not be found. I was assuming that if the package was
in the cached data, it is still going to be around. Fixed.