module-based test system, front end improvements, logging improvements
Moved the testing system to a module-based system rather than event-based.
This addresses a number of issues: 1) the dimsbuild system itself is module-
based, rather than event based, so this maintains a level of parallelism; 2)
fixes issues with events that were not named the same as the module they
came from for extension events; and 3) this shift allows for better integration
with an improved testing framework.  As such, the 'events' folder became the
'modules' folder, and several event folders were merged into their parent
module folder.

With the addition of 'runtest.py' at the testing root, all testing execution
has been removed from the individual modules themselves.  This allows a common
set of command line arguments to be available for any type of testing, whether
comprehensive or on a single module.  The way to run a test is to execute
runtest.py with the path to the python file to test.  This python file must
define a make_suite() function that returns a unittest.TestSuite instance with
all the tests to run.

Thus, to run all tests:
  `python runtest.py __init__.py`
To run just comps tests:
  `python runtest.py modules/comps/__init__.py`

This will likely change slightly in the future; support will probably be added
for directly naming the modules, rather than having to point to the actual file
in question; in this case, the above two cases would change to:

  `python runtest.py`
  `python runtest.py comps`

which is easier to type and remember.

A test log is now stored at the root of the test folder, 'test.log'.  It
contains the full log output of the most recent test; subsequent tests will
erase its content and overwrite it, so make sure to copy its contents if you
want to save it for any reason.

.hgignore
 * added patterns for the files generated by make
Makefile
 * changed paths from 'events' to 'modules'

test/__init__.py
 * removed all test execution and option parsing code (now in runtest.py)
 * added support for reading from a mainconfig file specified in options
 * added make_suite() to allow support for running via runtest.py

test/modules/*/__init__.py
 * main() and 'if __name__ == "__main__"' code blocks removed; execution is now
   handled via runtest.py
 * make_suite() changed to not accept a config parameter; config is now computed
   by the function itself
 * removed eventid, if present; since modules can contain more than one event,
   it doesn't make sense to define a single eventid

test/runtest.py
 * new execution wrapper for dimsbuild testing
 * supports a few command line arguments, see `runtest.py -h` for details
 * this file is responsible for running all tests; see description, above, for
   details on how to use it
 * --skip option doesn't currently work

test/modules/comps/__init__.py
 * fixed Test_Supplied so that it actually finds the correct file
test/modules/publish-setup/*
 * moved/merged into test/modules/publish
test/modules/ISO/*
 * moved/merged into test/modules/iso
test/modules/OS/*
 * moved/merged into test/modules/init
test/modules/setup/*
 * moved/merged into test/modules/init
test/modules/buildstamp/*
 * moved/merged into test/modules/infofiles
test/modules/discinfo/*
 * moved/merged into test/modules/infofiles
test/modules/treeinfo/*
 * moved/merged into test/modules/infofiles