test framework back end improvements (no event conversions)
Removed reliance on external config files for performing tests.  Modules can
now build their config entirely in memory.  This is good for a number of
reasons; no longer are we cluttering the test tree with extra config files,
now config and test cases are much closer together (the same file), and a
great deal of common setup can be factored using this method so that test
cases don't have to 'reinvent the wheel', so to speak, each time they want to
test something.

As a consequence of the above change, a lot of test cases ended up being a
single __init__.py file within the module's test folder.  Therefor, runtest.py
now has support for running python files directly from the modules folder in
addition to the normal support of <modid>/__init__.py.

runtest.py's command line parser was extended slightly to account for additional
ways to specify events.  Leaving the module to test blank will test all modules,
while specifying the module's id will test that specific module.  For example,
each of the following are equivalent

  # equivalence 1
  python runtest.py
  python runtest.py __init__.py

  # equivalence 2
  python runtest.py comps
  python runtest.py comps/__init__.py
  python runtest.py modules/comps/__init__.py

A config module was added to assist with the above task of building config
directly in memory.  Two functions are of particular note, make_default_config
and add_config_section.  The first creates a skeleton distro.conf file that
can be used in the majority of event test cases; it includes both 'main' and
'repos' top level elements, the second of which is populated with the 'base'
repository for 'fedora-6' by default.  This is sufficient for many events that
do not have their own config; however, it is clearly not enough to handle all
possible cases.  Thus, two other mechanisms exist to further modify this config
object.  The first was already mentioned above, the add_config_section function.
This function takes a config object and a string and attempts to parse the
string as XML and append it to the config object.  The second is via the _conf
attribute on EventTestCases.  This is a string or list of strings that will be
appended to the config file (using add_config_section) before being parsed.
In either case, any existing top-level elements with the same tag value will
be removed first.  Note that this is an in-place operation, meaning that any
changes to the config file will be reflected in all test cases that use it; be
careful to make as many copies of config as you need to prevent interference.

Added ModuleTestSuite, which is a simple subclass of unittest.TestSuite that
provides setUp and tearDown methods, much like unittest.TestCase already has.
This allows the tester to do certain things before and/or after a set of tests
has completed.  Currently, this is used to clean up the .dat autofiles used
by the rpm events to record their releases between sessions.

Event conversions will follow in a subsequent commit

dbtest/__init__.py
 * added ModuleTestSuite, which defines setUp and tearDown methods analagous
   to those in unittest.TestCase
 * EventTestCase.__init__ now does not require that the conf argument be
   specified.  If it is not, it makes a default config using config's
   make_default_config function.
 * updated make_suite to work with both file- and directory-based modules
dbtest/config.py
 * new module for constructing config objects in memory
 * see descrpition above and documentation inside for details
 * contains repo data for fedora 6, 7, and 8, centos 5, and redhat 5
dbtest/core.py
 * removed unnecessary __init__ methods
 * removed main block
 * make_extension_suite now automatically includes the tests from
   make_core_suite; as such, it is unnecessary to include both when making test
   cases
dbtest/mixins.py
 * removed unnecessary __init__ methods
 * made config optional on all mixin classes
dbtest/rpms.py
 * removed unnecessary __init__ methods
dbtest/runtest.py
 * improved command line arg parser; see description above