removed double-passing over options; removed cli support for events
Under the original dimsbuild design, events could provide and manipulate the
command line parser dimsbuild used for retrieving its arguments.  This resulted
in a complex, filtered two-layer parsing system to find and load the config
file, then find and load the appropriate dimsbuild modules, allow the events
to modify the cli parser, and finally parse the full set of command arguments.
Though this accomplished the goal, it was excessivly complex and prone to
error.  Due to the recent change of distro config file specification moving
from an option to an argument, the above system no longer works.

The new system is much simpler.  As none of the existing dimsbuild events made
use of the command line modification capabilities that were offered, this
feature was removed entirely.  Thus, command line parsing can occur in just
one pass, greatly simplifying the whole process.

A note on terminology: now that 'distro conf' has been removed from options
and added as an argument, we are choosing to instead call it a 'distribution
definition' or '.distro file'.  Meanwhile, what was formerly referred to as
the 'main config file' is now the only config file, thus just 'config file'.
This should help distinguish between the two types of files to prevent
confusion on the user end.

bin/dimsbuild
 * removed add_help_option=False from OptionParser creation; now that we use
   just a single pass, delaying help display is no longer required (this means
   that the -h output displays much faster when a config file is specified
   than it did before)
 * renamed '--dimsbuild-conf' option to '-c', '--config' - NOTE: this is
   different from what formerly occupied '-c', the distribution definition
 * argument parsing now performed in one pass instead of two
 * removed _filter_cmd_args
dimsbuild/main.py
 * removed calls to Event._add_cli and Event._apply_options; no longer
   supported
 * removed Build.apply_options; most of this code rolled into Build.__init__
dimsbuild/modules/core/installer/logos.py
 * fixed an error in verify_splash_exists