dimsbuild test back end changes - cleaner and crunchier for all!
This commit finally addresses the issue that has been present in testing for
a while - we really want to run core/extension tests on a test case of the
appropriate type, not some generic superclass.  The back end was changed to
allow for this; instead of passing an event id and a config, you pass an
EventTestCase subclass (the subclass, not an instance of the subclass) which
dbtest uses to perform tests.  Essentially, core.py now contains a number of
factory functions which decorate an instance of this class so that the
specific tests can be applied.  These changes largely remove the need to use
config.py to make config sections; instead, the superclass can simply define
_conf directly and automatically generate the correct config file.

dbtest/__init__.py
 * EventTestCase.__init__ no longer accepts an eventid; all subclasses are
   expected to define eventid and moduleid directly
 * fixed the code section that automatically added an appropriate config
   section to use moduleid instead of evenitd
 * EventTestCase.setUp now sets EventTestCase.event to None when the given
   event does not exist
 * added a pass definition for EventTestCase.runTest
 * added decorate, a function that takes a TestCase instance, the name of a
   method in that test case, and a pre and post function, returning that
   test case with the given method decorated by running the pre function
   before and post function after the method.
dbtest/core.py
 * changed make_core_suite and make_extension_suite to operate on
   EventTestCase classes rather than eventids
 * converted all EventTestCase classes into factory functions that transform
   an EventTestCase instance into an equivalent class
dbtest/mixins.py
 * changed fdm_make_suite and imm_make_suite to operate on EventTestCase
   classes rather than eventids
 * converted all EventTestCase classes into factory functions that transform
   an EventTestCase instance into an equivalent class
 * mixin classes no longer subclass EventTestCase
dbtest/modules/*
 * converted all event test cases to the new model
 * all test cases now define moduleid and eventid for the specific event they
   test
 * all modules now define an appropriate EventTestCase for each event they
   test
 * both 'publish' and 'release-files' still fail the first core event test;
   all other tests should succeed
dbtest/rpms.py
 * mixin classes no longer subclass EventTestCase
dimsbuild/modules/core/rpms/custom-repo.py
 * no longer setdefaults the 'repos' and 'source-repos' cvars