shared/repos: log ouput tweaks
Previously events that used the shared repos sync_repodata method printed
"Downloading repodata - '%s'" for each repository, whether new repodata
was downloaded or not.  This changeset moves the text from a
separate log call into the text argument of io.sync_input. In
this way, the text is only displayed if data is actually downloaded.