publish: workaround to get incremental sync working for publish
Most events deal with output cleanup via the FileIO.clean_eventcache
method. This method is not sufficient for the publish event, however,
since it stores content outside of its metadata folder.  To deal with
this, publish had been deleting all files in the publish folder
prior to hard linking updated content. In typical machine
configurations where the cache and the publish folder are on the
same partition, this can be accomplished with minimal performance
overhead.  When cache and publish folder are on separate partitions,
however sync copies the files rather than linking them.

The best fix for this issue would be to extend FileIO.clean_eventcache
and the difftest output handler support cleanup of output files outside
the metadata folder. This involves some fairly deep infrastructure work
as well as changes to many events.

This commit is a workaround that cobbles together strict syncing with
some manual cleanup activities to accomplish similar yet incomplete
results. The case that this workaround does not address is cleaning up
extraneous file in the publish dir during sessions when the publish
event does not run. This is not a huge issue, but it is an inconsistency
between publish and other events.