bug 147: allow file renaming using path elements
Added the ability to rename files using <path> style elements and their
derivatives.  Such elements now support a 'filename' attribute, which is
combined with the 'dest' attribut to form the full file name of the
destination file.

event.fileio was essentially rewritten in a large part in an attempt to make
things a little bit cleaner and simpler.  The API is mostly the same; the
only difference comes in adding files to by synced.  Instead of the single
method setup_sync, there are now two methods, add_xpath and add_fpath, the
first corresponding to xpath queries in the config file and the second to
paths on disk or on the web.  (There are also two convenience methods,
add_xpaths and add_fpaths, which can be used to add a list of xpaths or fpaths
one at a time).  These are in turn front ends for the 'real' method, add_item,
which takes a source and destination file ad adds it to the list of potential
sync targets.

share/spin/schemas/distro.conf/config-rpm.rng
 * added optional filename attribute to 'script' elements (this stuff should
   really be handled via a lib item now that it is somewhat prevalent)
spin/event/__init__.py
 * added dummy link arguments to Event.copy and Event.link; these don't
   actually do anything, but they make the signature for cache, copy, and
   link identical (helpful in fileio.py, below)
spin/event/fileio.py
 * major rewrite, see summary above for overview
 * removed IOObject.setup_sync
 * added IOObject.add_xpath, IOObject.add_xpaths for adding <path> elements by
   specifying an xpath query to them in the config file
 * added IOObject.add_fpath, IOObject.add_fpaths for adding existing files on
   the local filesytem or in a remote web location (any path supported by pps)
 * added IOObject.add_item, the back end to each of the above methods.  Takes
   a single src, dst pair, along with an optional id, mode, and/or prefix, and
   adds the information to the list of potential sync targets
 * cleaned up IOObject.sync_input significantly; sync method and callback are
   determined in the same place; actual sync logic is greatly simplified
 * removed IOObject.sync_items, IOObject.chmod_items
 * cleaned up IOObject.clean_eventcache somewhat; used to be performing
   findpaths on all listed outputs, which not only took significant amounts of
   time, but also allowed unexpected outputs to remain; I would like to fix
   this up even more, but can't afford the time at the moment
 * added IOObject._filter_data, a common method for processing 'what' arguments
 * added IOObject._process_path_xml, a common method for processing <path>
   elements and their derivatives
 * added TransactionData, a simple struct to keep track of source, destination,
   and mode for a transaction item (file)
spin/modules/*
 * all former calls to Event.io.setup_sync are now transformed to their
   appropriate counterpart, whether it be Event.io.add_fpath(s) or
   Event.io.add_xpath(s)
 * several modules had these sections optimized somewhat
spin/modules/core/rpms/custom-repo.py
 * added a mkdir on Event.auto_script.dirname prior to attempting to write
   script lines to it
spin/modules/shared/rpms.py
 * now makes use of Event.io._process_path_xml instead of separately handling
   <path> elements