further fix to relative path issues in fileio
Python's os library is somewhat limited in terms of the types of paths it
supports.  Specifically, use of os.PARDIR (..) causes issues with functions
such as os.path.exists, causing them to report back that the given item does
not exist even when performing the same call on the normalized path clearly
indicates that it does.  One solution would be to automatically normalize
paths as a part of most pps class methods; however, normpath can take a
significant amount of time, especially when called repeatedly in the process
of such tasks as findpaths and the like.  Thus, for now at least, it is up
to the user to ensure that all paths are normalized when necessary

spin/event/fileio.py
 * normalizes path before processing it in IOObject.add_item
spin/modules/shared/rpms.py
 * backed out changes of previous commit; superceded by above change