unified arch/basearch handling
Arch support has been clarified and standardized throughout spin.  We now
support two base archs, 'i386' and 'x86_64'; if no arch is given, 'i386' is
assumed.  In the background, when an arch is specified, spin remaps the given
arch to the highest arch in that arch's 'class'; 'athlon' for 'i386' and
'x86_64' for 'x86_64'.  This is primarily for yum's benefit so that it
considers all packages that potentially can match for a given arch when
performing depsolving. (That is to say, any time there are multiple archs
available for a given arch, spin will include all of them in the final output;
this is primarily an issue with the 'kernel' and 'glibc' packages.)

share/spin/schemas/distro.conf/main.rng
 * removed extra arch choices
spin/event/__init__.py
 * added ARCH_MAP, a simple dictionary mapping the supported base archs to
   the highest arch in the same arch class
 * Event.arch is now set by mapping the arch specified in the config file via
   ARCH_MAP, above.  Base arch is still computed the same way as before.
spin/modules/core/software/pkglist.py
 * removed special processing of arch; handled via standard Event.arch
 * simplified PkglistEvent.verify_kernel_arch; should now be much more
   easily extensible to any other arch we may choose to support.
spintest/runtest.py
 * added --clear-cache option
 * adjusted build root preservation to also respect the value of the
   --clear-cache option