bug 150 - allow module disables from config
The specfile and documentation allow for modules to be disabled from spin's
config file as well as from the distribution definition.  However, the actual
code hasn't supported this for some time.  This commit addresses this issue
by allowing modules to be disabled in the modules section of the spin config
file.

spin/event/loader.py
 * changed module loading behavior slightly:
   Old behavior:
    - do not load disabled default-on modules
    - load enabled default-off modules
   New behavior:
    - do not load disabled modules, default-on and default-off
    - load enabled default-off modules
   This means that being disabled has priority over being enabled.
spin/main.py
 * added code to support disabling modules from spin config file