idepsolver: smarter about packaged being removed
In IDespolver.setup(), the data from pkglist's previous run is read in
-- for each of the package tuples in the previous package list, the
package objects are found and saved for future use. If a package
object cannot be found for a package tuple, then the package is
removed. If a package object cannot be found for one of the deps of
that package, that dep is removed.

In effect, removing a package causes it and everything that requires
it to be depsolved. Removing just the dep and not the package
requiring it, in the case that the dep is missing, the other deps of
the package are not "lost." This makes depsolving to go faster.