idepsolver: made some more speed improvements
If a package that is in the package list created the last time
'pkglist' was run but is no longer there, that package and all the
packages it required were getting re-depsolved (if the package was
required this time). So in the case when a package was updated and the
older package was deleted from the repository that package and all its
dependencies were depsolved.

This has been changed now so that the deps are re-depsolved iff the
requires of the updated package has changed.