repos: added support for exclude-package in the repos element
Previously we supported exclude-package in the repo and comps elements.
This commit adds support in the repos element as well.  Packages listed
in this element are excluded both during comps file creation and from
pkglist dependency solving. We had discussed postponing this feature.
Since it turned out to be handy for troubleshooting broken package
dependencies in a repository, I went ahead. Changes include:

* added documentation for exclude-package in the repos container
* uncommented schema for exclude-package in the repos schema
* modified repos.py to add items from exclude-package elements to the
comps-excluded-packages and pkglist-excluded-packages cvars
* modified pkglist.py to include items from pkglist-excluded-packages in
the yumconf that is created prior to depsolving. Eventually we may need
to generalize the solution to allow for other yum global variables to be
passed from the repos event to the pkglist event.  We'll cross that
bridge when we come to it.