creating man pages using rst2man.py instead of docgen
We have moved our documentation files from docbook format to
reStructured Text format. With this change, we don't need the *.rng
files anymore because they were in docbook-format. I have converted
all of them to reST using "similar" directives. I still have to give
the .rst files a once over to make sure they are what we would expect
it to be.

I have also restructured the docsrc/ folder. The docsrc folder has a
man/ folder (in preparation for an html/ folder parallel to it). The
man/ folder has three folders: distro/, spin/, and tools/. The distro/
folder has the reST files related to the distro man page. The spin/
folder has the reST files related to the spin.conf man page. The
tools/ folder has the tools needed to convert reST files to man files.