Fixed Bug 165.
If there is an update available for a package, a series of checks are
performed. Before this commit these checks were as follows:
1. Check all the packages that require the package with the available
   update and keep a record of what exactly it is about the package
   that it requires.
2. Then make sure that the update package provides everything the
   older package did.
3. If these checks are passed, replace the old package with the new
   one.

There was one issue with this which is, there were some packages that
required (<package name>, None, (None, None, None)), i.e. any package
that provides <package name>. The aforementioned series of checks
didn't always catch this because there were some packages,
e.g. 'syslinux' that only provided a specific version of itself. To
get around this, some more checks have been added.

In the following lines, package A refers to a package that requires
package B, and package B has an update available called package C.

1. If package A doesn't require an explicit version of package B,
   replace package B with package C.
2. If package A requires a version {<=,<,==,>,>=} that of package B
   and package C also satisfies this requirement, replace package B
   with package C.