Added support for images with gradients.
The ImageGradient class has support for creating horizontal gradients
on images. It has, currently, one method of interest: draw_gradient(),
which takes in two parameters: the start_color and the
end_color. Based on these two colors, a gradient is drawn on the image
with the start_color on the "left side" of the image, and the
end_color on the "right side" of the image.

Modified the LogosHandler class to first look in the distro-specific
folder for an image. If the image is found there, copy it to the build
folder. If the image is not found there, use the fallback image (found
in the currently empty fallback/ folder) and an ImageGradient object
to create an image and save it in the build folder.