locals/logos-rpm: updated locals to reflect changed in event
The locals dictionary for the logos-rpm mapped a file to its XWindow
type. For example, if a file is required by GNOME only, the file's
locals entry would have a key: mapping pair of "/path/to/file":
'gnome'. With changes to the logos-rpm and its shared/ folder, we
don't need to store the XWindow type in the locals anymore.