logos-rpm: all user-specified files used
If a user specifies a folder to use as input to the logos-rpm event,
use all the files in that folder. If, for any reason, the files in the
shared/distros, shared/fallback, and shared/common directories are
used, then the file is used only if it is present in the logos-rpm
event's locals dictionary. This allows us to have files based on
anaconda versions.