The Repo.__str__() method writes local URLs to 'baseurl' key.
The Repo.__str__() method was (incorrectly) writing the remote URL as
the value of the 'baseurl' option. This was causing the 'pkglist'
event's setup() step to download the repodata folder from the remote
location which is (usually) time-consuming.

To get around this problem, the Repo class now has a tostring() method
that accepts an optional 'remote' parameter. If 'remote' is set to
True, the 'baseurl' option's value is the remote URL; if 'remote' is
set to False, the 'baseurl' option's value is the local URL. The
Repo.__str__() method makes use of the Repo.tostring() method and sets
'remote' to be False. This speeds up the pkglist's setup() step by some.