release-rpm: The .repo file generated has the remote URL for input repos
With changes to the Repo class, the Repo.__str__() method always
returns the local (cached) URL as the baseurl. This is not what the
baseurl value should be in the .repo file installed on the end-user
machine. The baseurl in that case should be the remote (input)
URL.

Therefore, when the input repos are being written to the .repo file,
the Repo.tostring() method is called directly with the 'remote'
parameter set to True.