rpm: if version doesn't contain an int, use the version string
For determining custom rpm release-versions, we attempt to use
an integer found in the user-specified version string for interop with
3rd party repofiles that use the $releasever macro.  Previously, if the
string contained no integer, we raised an exception. Now we just
use the string. This allows specifying a version of 'development' when
creating distros based off the fedora development repository, for
example.