locals/logos_rpm: added distro-specific locals dictionary
The L_LOGOS_RPM_INFO locals dictionary contains distro-specific
information: the scriptlets to be installed for that distro, the
shared folder to look in for the files to be installed by the RPM, the
colors to use for the images we create etc.

Now that this information is in a locals dictionary, we can get rid of
all the conditional logic in the logos RPM event.