Fixed bug 173.
In RepoEventMixin.read_repodata(), when Repo.read_repomd() is called a
pps.path.error.PathError exception is raised if the repomd.xml file is
missing. If this happens, it means that the repodata/ folder is
missing and we move on. In the case of the 'sources' event, it is
completely valid for the repodata/ folder to be missing. In the case
of 'repos' event, there is a verify_repodata_exists() method that
makes sure that the repodata/ folder exists. So, it is fine to move on
in the RepoEventMixin.read_repodata() method. Moving on also causes
the Repo.datafiles attribute to not get populated.

If the repodata/ folder is missing, the Repo.datafiles attribute is
not set (as explained above), so when we try to get the primary.xml.gz
file's location a KeyError is expected and is passed. This results in
the Repo.pkgsfile not being created.