Updated references to classes that are in xmllib.{macros,errors}
The xmllib.errors modules contains xmllib-related errors. Most of the
items in this module used to be in xmllib.tree. All the files using
classes and functions in xmllib.errors have been updated.

The shared/bootcfg.py file uses the Macro class which used to be in
xmllib.config but is now in xmllib.macros.