logos-rpm: creating default images if no images found in shared paths
The FallbackHandler and DistroSpecificFileHandler have been combined
to form the DistroFilesHandler class. The DistroSpecificHandler's
generate() method first copies all the images from the shared paths to
the build folder. Once that is done, it checks to make sure that all
the images that are supposed to be there are in fact present, and if
not an image is created.

The locals dictionary L_LOGOS_RPM_FILES has been updated to now have
the 'xwindow_type', 'image_width', 'image_height', and (optionally)
'image_format' keys.