'logos' event does the "right" thing in the case of missing logos RPM
If the logos-rpm event is disabled, and there are no logos RPMs in the
list of downloaded packages, then the expected behavior is that the
splash image from the input repository is downloaded to repository
composed by Spin.

To accomplish this, the 'isolinux' event downloads the splash image
from the input repository unless the 'installer-splash' control
variable is set by the 'logos' event. If 'installer-splash' is set,
then that file is downloaded instead of the input repository's splash
image.

The diskboot image had to be modified because the
'isolinux-files/installer-splash' control variable doesn't exist
anymore.