logos-rpm: computing the index of the "white" color programmatically
Previously, it was a requirement of the logos-rpm event to have a
white color (in a limited color palette image) to be at the same
index. Now, we figure out the index of the white color automagically
and use it. This way, the requirement is reduced to just having the
color white in the palette. Much nicer this way.