Gotten rid of the 'ImageMagick' dependency!
The only place we were using something from the "ImageMagick" package
was in the logos-rpm event to convert a png file to an xpm file. GRUB
uses this xpm file as the background of the screen when the GRUB menu
is displayed.

Before, converting the image to an xpm, we were writing some text to
it. We don't write any text to the GRUB background image anymore, so
we can have a static xpm file in the spin-logos-rpm RPM. If the
spin-logos-rpm is not installed on the machine running Spin, then the
GRUB menu has no background (i.e. a black background).

Made all the necessary changes in the sources to remove references to
ImageMagick.