idepsolver: the "indentation error" fix a few days ago broke things.
The indentation error that I was seeing a few days ago was happening
because the package to be removed was not there in the list of cached
items; I incorrectly assumed it to be an indentation error, and
"fixed" it. This commit makes the necessary changes to that the
removal of packages has the desired effect.