Not converting trigger*_text to trigger*_script tags.
With the changes in `rendition.mkrpm', trigger*_script[1] tags are no
longer supported. Made changes in the logos-rpm event to pass
trigger*_scripts[1] parameters.

Also, removing the conversion of trigger*_text to trigger*_script, now
that trigger*_script doesn't exist (replaced by trigger*_scripts),
because:

1. It is difficult to know which script file out of the list of
   trigger*_scripts, the text should be added to.

2. If a dummy trigger*_script file is created and added to
   trigger*_scripts, it is difficult to compute a unique name. Even if
   it were easy, it is difficult to compute the index where this dummy
   trigger*_script file should be inserted.

Therefore, it is left to the client to figure all this out.


[1] triggerin_script(s), triggerun_script(s), triggerpostun_script(s)