Reorganized the triggers in L_LOGOS_RPM_DISTRO_INFO.
Previously, the 'triggerin' and 'triggerun' keys were defined
independently, with its children being a list of things to trigger
on. Now, there is 'triggers' key which defines a list of items to
trigger on, with its children being 'triggerin', 'triggerun', and,
potentially in the future, 'triggerin_flags', 'triggerun_flags' etc.

Updated the logos-rpm module to reflect these changes.